Move app bootstrap code to module
