+++ config.toml.dist
@@ -1,4 +1,4 @@
-address = "127.0.0.1"
+ip = "127.0.0.1"
 port = 80
 
 [[servers]]
